requirements.txt
asgiref==3.11.1
async-timeout==5.0.1
Django==5.2.12
django-filter==25.2
django-redis==6.0.0
djangorestframework==3.16.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
redis==7.3.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
django-cors-headers
Pillow>=10.0.0
djangorestframework-simplejwt
requests
python-dotenv